fix: move getSortedReportActionsForDisplay to Utils and use in iouTransaction
